Rename `StypRuleGetter` to `RefStypRule`
