+++ src/rule/index.ts
@@ -1,6 +1,5 @@
 export * from './properties';
 export * from './root';
 export * from './rule';
-export * from './rule-getter';
 export * from './rule-ref';
 export * from './rules';
